+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `GradOf` constructors: `green`, `greenAccent`; `lightGreen`,
   `lighGreenAccent`; `lime`, `limeAccent`; `teal`, `tealAccent`.
 
+[Indigo gradients](https://github.com/dartoos-dev/eo_color/issues/128):
+
+- `GradOf` constructors: `indigo`, `indigoAccent`.
+
 ## [2.1.0] - 2021-11-08
 
 ### Added